refactor: split oversized files (ExecuteQueryUseCase, workspaces routes, RouteMessage, injection)
